requirements.txt
asgiref==3.5.2
asttokens==2.0.5
Babel==2.10.1
backcall==0.2.0
certifi==2022.5.18.1
colorama==0.4.4
cssselect2==0.6.0
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-oscar-api==3.1.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
easy-thumbnails==2.8.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.12.1
ipython==8.4.0
jedi==0.18.1
lxml==4.9.0
matplotlib-inline==0.1.3
parso==0.8.3
phonenumbers==8.12.49
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
purl==1.6
pycountry==22.3.5
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.10
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
stack-data==0.2.0
svglib==1.3.0
tinycss2==1.1.1
traitlets==5.2.2.post1
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
